+++ test/controllers/auth_controller_test.rb
@@ -2,7 +2,7 @@
 
 require "test_helper"
 
-class AccessControlControllerTest < ActionDispatch::IntegrationTest
+class AuthControllerTest < ActionDispatch::IntegrationTest
   def setup
     # テスト用の環境変数を設定
     ENV["ALLOWED_EMAIL_ADDRESSES"] = "okita2710@gmail.com"
@@ -27,7 +27,7 @@ class AccessControlControllerTest < ActionDispatch::IntegrationTest
 
   test "許可されたメールアドレスの認証" do
     post authenticate_email_path, params: { email: "okita2710@gmail.com" }
-    assert_redirected_to verify_code_path
+    assert_redirected_to verify_access_code_path
     assert_equal "okita2710@gmail.com", session[:pending_email]
     assert_equal 1, ActionMailer::Base.deliveries.length
   end
@@ -41,11 +41,11 @@ class AccessControlControllerTest < ActionDispatch::IntegrationTest
 
   test "正しい認証コードの検証" do
     post authenticate_email_path, params: { email: "okita2710@gmail.com" }
-    assert_redirected_to verify_code_path
+    assert_redirected_to verify_access_code_path
 
     code = EmailVerificationCode.last.code
 
-    post verify_code_path, params: { code: code }
+    post verify_access_code_path, params: { code: code }
     assert_redirected_to "/home"
     assert session[:email_authenticated]
     assert_equal "okita2710@gmail.com", session[:authenticated_email]
@@ -55,10 +55,10 @@ class AccessControlControllerTest < ActionDispatch::IntegrationTest
 
   test "メール認証済み時のホームページリダイレクト" do
     post authenticate_email_path, params: { email: "okita2710@gmail.com" }
-    assert_redirected_to verify_code_path
+    assert_redirected_to verify_access_code_path
 
     code = EmailVerificationCode.last.code
-    post verify_code_path, params: { code: code }
+    post verify_access_code_path, params: { code: code }
     assert_redirected_to "/home"
 
     get root_url
@@ -82,9 +82,9 @@ class AccessControlControllerTest < ActionDispatch::IntegrationTest
 
   test "間違った認証コードの検証失敗" do
     post authenticate_email_path, params: { email: "okita2710@gmail.com" }
-    assert_redirected_to verify_code_path
+    assert_redirected_to verify_access_code_path
 
-    post verify_code_path, params: { code: "000000" }
+    post verify_access_code_path, params: { code: "000000" }
     assert_response :success
     assert_nil session[:email_authenticated]
     assert_not_nil session[:pending_email]
@@ -92,12 +92,12 @@ class AccessControlControllerTest < ActionDispatch::IntegrationTest
 
   test "期限切れ認証コードの検証失敗" do
     post authenticate_email_path, params: { email: "okita2710@gmail.com" }
-    assert_redirected_to verify_code_path
+    assert_redirected_to verify_access_code_path
 
     code_record = EmailVerificationCode.last
     code_record.update!(expires_at: 1.minute.ago)
 
-    post verify_code_path, params: { code: code_record.code }
+    post verify_access_code_path, params: { code: code_record.code }
     assert_response :success
     assert_nil session[:email_authenticated]
     assert_not_nil session[:pending_email]
@@ -105,16 +105,16 @@ class AccessControlControllerTest < ActionDispatch::IntegrationTest
 
   test "空の認証コードのエラー表示" do
     post authenticate_email_path, params: { email: "okita2710@gmail.com" }
-    assert_redirected_to verify_code_path
+    assert_redirected_to verify_access_code_path
 
-    post verify_code_path, params: { code: "" }
+    post verify_access_code_path, params: { code: "" }
     assert_response :success
     assert_nil session[:email_authenticated]
     assert_not_nil session[:pending_email]
   end
 
   test "保留メールなしでのルートページリダイレクト" do
-    get verify_code_get_path
+    get verify_access_code_get_path
     assert_redirected_to root_path
   end